requirements.txt
crewai>=0.28.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
websockets>=11.0.3
jinja2>=3.1.2
aiofiles>=23.2.1
pandas>=2.1.1
numpy>=1.26.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pymongo>=4.5.0
python-dateutil>=2.8.2
langchain>=0.0.335
langchain-openai>=0.0.2
openai>=1.3.0
